+++ modules/gdnative/nativescript/nativescript.h
@@ -371,9 +371,9 @@ public:
 	virtual int profiling_get_frame_data(ProfilingInfo *p_info_arr, int p_info_max);
 
 	int register_binding_functions(godot_instance_binding_functions p_binding_functions);
-	void unregister_binding_functions(size_t p_idx);
+	void unregister_binding_functions(int p_idx);
 
-	void *get_instance_binding_data(size_t p_idx, Object *p_object);
+	void *get_instance_binding_data(int p_idx, Object *p_object);
 
 	virtual void *alloc_instance_binding_data(Object *p_object);
 	virtual void free_instance_binding_data(void *p_data);
